New option to allow override of how fstab entries get created.
New option to override how fstab entries get created. For vols such as vxvm,
using uuid does not work. it should always be devpath. This will allow
to control this.
ncm-lib-blockdevices code has been updated to check for
this option, if defined and true, it will ignore uuid and
always use devpath when creating entries in fstab.
Requires ncm-lib-blockdevices >=25.x